fix(home): extract LoadingSkeleton to its own file
The client infinite-scroll component imported LoadingSkeleton from
component.tsx (a server component). Turbopack traced the import and
tried to bundle component.tsx's server-only transitive deps
(@opentelemetry/context-async-hooks pulls node:async_hooks) into the
client bundle, breaking the build with 'Module not found: async_hooks'.

Move LoadingSkeleton into its own module with no server-only imports
so either side can consume it.
