+++ apps/web/app/home/component.tsx
@@ -2,12 +2,11 @@ import * as React from 'react';
 
 import { SidebarCounts } from '~/core/io/fetch-sidebar-counts';
 
-import { Skeleton } from '~/design-system/skeleton';
-
 import {
   type GovernanceHomeReviewCategory,
   type GovernanceHomeStatusFilter,
 } from './fetch-active-proposals-in-editor-spaces';
+import { LoadingSkeleton } from './loading-skeleton';
 import { HomeProposalsInfiniteScroll } from './home-proposals-infinite-scroll';
 import { MyGovernanceProposalsList } from './my-governance-proposals-list';
 import { PendingProposalsPage } from './pending-proposals-page';
@@ -99,18 +98,6 @@ export async function Component({
         />
       </div>
     </>
-  );
-}
-
-export function LoadingSkeleton() {
-  return (
-    <div className="space-y-4 rounded-lg border border-grey-02 p-4">
-      <div className="space-y-2">
-        <Skeleton className="h-5 w-36" />
-        <Skeleton className="h-4 w-20" />
-      </div>
-      <Skeleton className="h-5 w-48" />
-    </div>
   );
 }
