+++ apps/web/app/home/loading-skeleton.tsx
@@ -1,0 +1,13 @@
+import { Skeleton } from '~/design-system/skeleton';
+
+export function LoadingSkeleton() {
+  return (
+    <div className="space-y-4 rounded-lg border border-grey-02 p-4">
+      <div className="space-y-2">
+        <Skeleton className="h-5 w-36" />
+        <Skeleton className="h-4 w-20" />
+      </div>
+      <Skeleton className="h-5 w-48" />
+    </div>
+  );
+}
